step_add function use C

+++ step.c
@@ -3,10 +3,10 @@
 //It's have a CPlusPlus edition
 
 
-long double step_add(long double max)  //max must be float
+long step_add(long max)  //max must be float
 {	
- 	long double j=0.0; //define j
- 	for (long double i=0.0;i<max+1;i+=1)  //a easy for loop
+ 	long j=0.0; //define j
+ 	for (long i=0.0;i<max+1;i+=1)  //a easy for loop
  	{
  		j+=i;  //Plus
  	}
@@ -18,7 +18,6 @@ long double step_add(long double max)  //max must be float
 
 int main()
 {
-
 
 	return 0;
